% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubicar.R
\name{cubicar}
\alias{cubicar}
\title{Cubicar un numero}
\usage{
cubicar(x)
}
\arguments{
\item{x}{numeric(1) Any numeric value}
}
\value{
un numero cubico
}
\description{
Revisa que la entrada sea numerica
}
\examples{
cubicar(2)
}
